Changed character stamina, added default character, added health/stamina pickups
Halved the amount of stamina regenerated from 0.05 to 0.025.
Made it so if a player doesn't pick a character, the Knight is selected.

Added full functionality for health/stamina pickups.
Picking up health regenerates 33.3% of a character's maximum health
Picking up stamina regenerates 20% of a character's maximum stamina.
Both pickups have a cooldown (set in inspector).
While on cooldown pickup is not visible.